modified readme

+++ meme/views.py
@@ -106,9 +106,11 @@ def remove_post(request, post_id):
         data = json.loads(request.body)
         post_id = data.get("post_id")
         post = Post.objects.get(pk=post_id)
-        post.delete()
-        return JsonResponse({"success": True})
-
+        if request.user.profile == post.author:
+            post.delete()
+            return JsonResponse({"success": True})
+        else:
+            return JsonResponse({"success": False})
 
 def create_post(request):
     if request.method == "POST":
